requirements.txt
Pillow>=9.0.1
pytesseract>=0.3.10
tqdm>=4.64.0
numpy>=1.21.5